requirements.txt
flask==2.2.5
boto3==1.18.47
kafka-python